refactor: storage 유틸 함수 수정
