+++ frontend/src/utils/index.js
@@ -1,3 +1,3 @@
 export { getId, getAvatarKey } from './id';
 export { httpRequest } from './httpRequest';
-export { getLocalStorage, setLocalStorage, removeLocalStorage } from './localStorage';
+export { storage } from './storage';
